+++ edit/src/class_instruction.cpp
@@ -27,7 +27,7 @@ void editor::instruction::process_special_chars() {
 
 }
 
-bool editor::instruction::set_replace_instruction(int start, string text) {
+bool editor::instruction::set_replace_instruction(long long int start, string text) {
 	/***
 	bool set_replace_instruction():
 		- Sets the current instruction to type "replace"
@@ -64,7 +64,7 @@ bool editor::instruction::set_replace_instruction(int start, string text) {
 	return true;
 }
 
-bool editor::instruction::set_insert_instruction(int start, string text) {
+bool editor::instruction::set_insert_instruction(long long int start, string text) {
 	/***
 	bool set_insert_instruction():
 		- Sets the current instruction to type "insert"
@@ -101,7 +101,7 @@ bool editor::instruction::set_insert_instruction(int start, string text) {
 	return true;
 }
 
-bool editor::instruction::set_remove_instruction(int start, int end) {
+bool editor::instruction::set_remove_instruction(long long int start, long long int end) {
 	/***
 	bool set_remove_instruction():
 		- Sets the current instruction to type "remove"
@@ -168,11 +168,11 @@ int editor::instruction::get_operation_type() {
 	return operation;
 }
 
-int editor::instruction::get_start_position() {
+long long int editor::instruction::get_start_position() {
 	return start_position;
 }
 
-int editor::instruction::get_end_position() {
+long long int editor::instruction::get_end_position() {
 	return end_position;
 }
 
@@ -180,7 +180,7 @@ string editor::instruction::get_text() {
 	return text_input;
 }
 
-editor::instruction create_replace_instruction(int start_position, string text) {
+editor::instruction create_replace_instruction(long long int start_position, string text) {
 	editor::instruction new_instruction;
 	
 	new_instruction.set_replace_instruction(start_position, text);
@@ -188,7 +188,7 @@ editor::instruction create_replace_instruction(int start_position, string text) 
 	return new_instruction;
 }
 
-editor::instruction create_insert_instruction(int start_position, string text) {
+editor::instruction create_insert_instruction(long long int start_position, string text) {
 	editor::instruction new_instruction;
 	
 	new_instruction.set_insert_instruction(start_position, text);
@@ -196,7 +196,7 @@ editor::instruction create_insert_instruction(int start_position, string text) {
 	return new_instruction;
 }
 
-editor::instruction create_remove_instruction(int start_position, int end_position) {
+editor::instruction create_remove_instruction(long long int start_position, int end_position) {
 	editor::instruction new_instruction;
 	
 	new_instruction.set_remove_instruction(start_position, end_position);
@@ -314,21 +314,21 @@ editor::instruction parse_instruction_string(string instruction_string) {
 	Now, to actually create the 'instruction' object
 	*/
 	if (is_replace_instruction) {
-		int start_position = (int)stol(instruction_vector[1]);
+		long long int start_position = (long long int)stol(instruction_vector[1]);
 		
 		return create_replace_instruction(start_position, instruction_vector[2]);
 	}
 	
 	if (is_insert_instruction) {
-		int start_position = (int)stol(instruction_vector[1]);
+		long long int start_position = (long long int)stol(instruction_vector[1]);
 		
 		return create_insert_instruction(start_position, instruction_vector[2]);
 	}
 	
 	if (is_remove_instruction) {
-		int start_position = (int)stol(instruction_vector[1]);
+		long long int start_position = (long long int)stol(instruction_vector[1]);
 		
-		int end_position = (int)stol(instruction_vector[2]);
+		long long int end_position = (long long int)stol(instruction_vector[2]);
 		
 		return create_remove_instruction(start_position, end_position);
 	}
